+++ app/services/unfavourite_service.rb
@@ -2,6 +2,10 @@
 
 class UnfavouriteService < BaseService
   def call(account, status)
+    # attempt to unfavourte, responds to validator
+    unfav = Unfavourite.create!(account: account, status: status)
+    raise ActiveRecord::RecordInvalid unless unfav.valid?
+
     favourite = Favourite.find_by!(account: account, status: status)
     favourite.destroy!
     favourite
